% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_informations.R
\name{add_meta}
\alias{add_meta}
\title{Add new group for spatial spots}
\usage{
add_meta(add_meta, dirs = "shinyspatial_app", colname = "new_group")
}
\arguments{
\item{add_meta}{new group of spots, characters or data frame/data table}

\item{dirs}{directory shinyapp}

\item{colname}{group names, if add_meta is the data frame, colname could be used to identify the column of add_meta}
}
\value{
regenerate the meta.Rds and meta_group.Rds with new groups
}
\description{
add some information into meta.Rds and meta_group.Rds
}
