% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolusiondwls.R
\name{DWLSmatrix}
\alias{DWLSmatrix}
\title{get DWLS signature gene-cell matrix for deconvolusion}
\usage{
DWLSmatrix(matrix, sign_gene, cell_cols, meta)
}
\arguments{
\item{matrix}{normalized scRNA-seq matrix}

\item{sign_gene}{signature gene}

\item{cell_cols}{colums of cell type}

\item{meta}{scRNA-seq meta}
}
\description{
get DWLS signature gene-cell matrix for deconvolusion
}
