% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwlsfun.R
\name{enrich_deconvolution}
\alias{enrich_deconvolution}
\title{function of Giotto DWLS deconvolusion}
\usage{
enrich_deconvolution(expr, log_expr, cluster_info, ct_exp, cutoff)
}
\description{
function of Giotto DWLS deconvolusion
}
