% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_code_server.R
\name{lib_server}
\alias{lib_server}
\title{server code prepared for shinySRT}
\usage{
lib_server()
}
\arguments{
\item{shiny.dir}{specify directory to create the shiny app in. Default is
to create a new directory named "shinyspatial_app"}
}
\value{
server code required for shiny app
}
\description{
Prepare the server code for shinyApp.
}
