% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_process.R
\name{load_spatial}
\alias{load_spatial}
\title{Input the SRT data}
\usage{
load_spatial(
  matx = NULL,
  meta = NULL,
  coordi = NULL,
  image = NULL,
  species = "hg",
  x_reverse = F,
  min_cells = 3,
  scale_factors_file = NULL
)
}
\arguments{
\item{matx}{gene-barcode matrix file path, formation:.txt, .csv, .tsv, .xlsx}

\item{meta}{meta data file path, formation:.txt, .csv, .tsv, .xlsx}

\item{coordi}{file path, Spatial location information, formation:.txt, .csv, .tsv, .xlsx}

\item{image}{image file path, formation: .png, .jepg, .tiff.}
}
\value{
object for process
}
\description{
iuput SRT matrix to generate a seurat obj or SRT list
}
