% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_process.R
\name{multi_dir_spatial}
\alias{multi_dir_spatial}
\title{multi-dir generate obj for shiny app}
\usage{
multi_dir_spatial(
  multi_dir,
  species = "hg",
  x_reverse = F,
  min_cells = 3,
  resolution = 1,
  npcs = 20
)
}
\arguments{
\item{multi_dir}{directory of multi SRT data directory}
}
\value{
a object for process
}
\description{
multi-dir generate obj for shiny app
}
