% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_process.R
\name{obj_list_process}
\alias{obj_list_process}
\title{Seurat process for multi-object(object in list)}
\usage{
obj_list_process(obj_list, species = "hg", npcs = 20, resolution = 1)
}
\arguments{
\item{obj_list}{seurat object list (None-processing)}
}
\value{
a object for shinySRT
}
\description{
Traditional seurat process with spatial data (using SCT)
}
