% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwlsfun.R
\name{page_dt_method}
\alias{page_dt_method}
\title{enrich}
\usage{
page_dt_method(
  sign_matrix,
  expr_values,
  min_overlap_genes = 5,
  logbase = 2,
  reverse_log_scale = TRUE,
  output_enrichment = c("original", "zscore"),
  p_value = FALSE,
  include_depletion = FALSE,
  n_times = 1000,
  max_block = 2e+07,
  verbose = TRUE
)
}
\description{
enrich
}
