% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_file.R
\name{preparedata_web}
\alias{preparedata_web}
\title{Prepare data files required for shiny app(web)}
\usage{
preparedata_web(
  dat,
  meta.to.include = NA,
  tooltip_col = NULL,
  maxlevel = 50,
  shiny.dir = "shinyspatial_app",
  chunkSize = 500,
  gex.assay = NA,
  gex.slot = c("data", "scale.data", "counts"),
  gene.mapping = F,
  default.gene1 = NA,
  default.gene2 = NA,
  default.multigene = NA
)
}
\arguments{
\item{dat}{imported data for shinySRT}
}
\description{
This function could generate data files for shiny app by saptial transcriptomic data, e.g. matrix, metadata, coordination, H&E images.
}
\details{
Generate data files required for shiny app.Six files will be generated,
(1) shinySRT config \code{meta_group.Rds}
(2) the gene mapping object config \code{genesets.Rds}
(3) the gene expression \code{data.h5}
(4) the spatial metadata \code{meta.Rds}
(5) the defaults for the shiny app \code{df_select.Rds}
(6) the spatial image \code{image.Rds}
Note that both \code{preparedata_shinyspatial} and \code{prepare_code} functions are ran when
running the wrapper function \code{CreateshinySRT}.
}
