% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolusiondwls.R
\name{runDWLSDeconv}
\alias{runDWLSDeconv}
\title{Deconvolusion of Giotto in shinySRT}
\usage{
runDWLSDeconv(
  norm_mtx,
  logbase = 2,
  cell_metadata,
  cluster_column,
  sign_matrix,
  n_cell = 50,
  cutoff = 2,
  name = NULL
)
}
\arguments{
\item{norm_mtx}{normalized data matrix (or sparse)}

\item{logbase}{base used for log normalization}

\item{cell_metadata}{spatial meta}

\item{cluster_column}{name of cluster column}

\item{sign_matrix}{signature gene-cell matrix for deconvolusion}

\item{number}{of cells per spot}
}
\value{
data.table of cellular component
}
\description{
This function could be used to predict the cellular component of each spots, refering to Giotto's DWLS deconvolusion function.
}
