% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolusiondwls.R
\name{scran_hvg}
\alias{scran_hvg}
\title{Find highly variable gene}
\usage{
scran_hvg(
  mtx,
  meta,
  colcluster,
  method = "scran",
  pval = 0.01,
  logFC = 0.5,
  min_feats = 3
)
}
\arguments{
\item{pamtx}{normalized data matrix}
}
\description{
Find highly variable gene
}
