% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolusiondwls.R
\name{scran_norm}
\alias{scran_norm}
\title{scran normalization}
\usage{
scran_norm(
  mymatrix,
  log_offset = 1,
  logbase = 2,
  scalefactor = 6000,
  library_size_norm = TRUE,
  log_norm = TRUE
)
}
\arguments{
\item{mymatrix}{transcriptomic data matrix}

\item{log_offset}{offset}

\item{logbase}{base used for log normalization}
}
\description{
scran normalization
}
