% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_process.R
\name{seurat_sp_process}
\alias{seurat_sp_process}
\title{Seurat process}
\usage{
seurat_sp_process(obj, species = "hg", npcs = 20, resolution = 1)
}
\arguments{
\item{obj}{seurat object (None-processing)}

\item{npcs}{Dimensions used of pca}

\item{resolution}{precision of clustering}
}
\value{
a object for shinySRT
}
\description{
Traditional seurat process with spatial data (using SCT)
}
