% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_process.R
\name{single_op_dir}
\alias{single_op_dir}
\title{Load the SRT data from dir and process}
\usage{
single_op_dir(
  dir,
  species = "hg",
  x_reverse = F,
  min_cells = 3,
  resolution = 1,
  npcs = 20
)
}
\arguments{
\item{dir}{directory of SRT data, matrix, position information, image, et.al.}
}
\value{
a object for shinySRT
}
\description{
Load the SRT data from dir and process
}
