% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_process.R
\name{single_op_file}
\alias{single_op_file}
\title{Load the SRT data  and process}
\usage{
single_op_file(
  matx = NULL,
  meta = NULL,
  coordi = NULL,
  image = NULL,
  species = "hg",
  x_reverse = F,
  min_cells = 3,
  resolution = 1,
  scale_factors_file = NULL,
  npcs = 20
)
}
\arguments{
\item{matx}{gene-barcode matrix file path, formation:.txt, .csv, .tsv, .xlsx}

\item{meta}{meta data file path, formation:.txt, .csv, .tsv, .xlsx}

\item{coordi}{file path, Spatial location information, formation:.txt, .csv, .tsv, .xlsx}

\item{image}{image file path, formation: .png, .jepg, .tiff.}
}
\value{
a object for shinySRT
}
\description{
Load the SRT data  and process
}
