% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_process.R
\name{spatial_load_dir}
\alias{spatial_load_dir}
\title{Load the SRT data from dir}
\usage{
spatial_load_dir(x, species = "hg", x_reverse = F, min_cells = 3, ...)
}
\arguments{
\item{x}{directory of SRT data, matrix, position information, image, et.al.}
}
\value{
a object for process
}
\description{
Load the SRT data from dir
}
