% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_code_ui.R
\name{ui_load}
\alias{ui_load}
\title{ui code prepared for shinyspatial}
\usage{
ui_load()
}
\arguments{
\item{df_select}{default information to provide the slice number


ui_load : load package in ui scirpts
ui_head : with web's header and menu}
}
\value{
data files required for shiny app
}
\description{
ui code prepared for shinyspatial
}
